+++ app/new-bookshop/webapp/i18n/i18n_en_US.properties
@@ -33,17 +33,42 @@ tableNameColumnRating=Rating
 buttonCreate=Create
 
 #~~~ Object View ~~~~~~~~~~~~~~~~~~~~~~~~~~
-#XTIT: Object view title
-objectViewTitle=<Books> Details
 
-#XTIT: Object page title
-objectTitle=<Books>
+#XBUT: text for button
+buttonEdit=Edit
+buttonDelete=Delete
 
-#XTIT: Label for the ID
-IDLabel=ID
+#XFLD: The label of the header of Object Page Layout
+labelPrice=Price:
+labelStock=Stock:
 
-#XTIT: Label for the 
-Label=
+#XTIT: The title of the Object Page Section
+titleFullDescription=Full description
+titleDescription=Description
+titleGenre=Genre
+titleRating=Rating
+titleOrders=Orders
+
+#XCOL: The name columns of the Responsive Table
+tableNameColumnCustomerName=Customer name
+tableNameColumnDeliveryDate=Delivery date
+tableNameColumnPhoneNumber =Phone number
+tableNameColumnCount       =Count
+
+#XTIT: The title of the grid table
+gridTableTitle=Orders
+
+#XBUT: text for button(delete orders)
+buttonGridTableDelete=Delete
+
+#The name columns of the Grid Table 
+gridTableNameColumnCustomerName=Customer name
+gridTableNameColumnDeliveryDate=Delivery date
+gridTableNameColumnPhoneNumber =Phone number
+gridTableNameColumnCount       =Count
+
+
+
 
 #~~~ Not Found View ~~~~~~~~~~~~~~~~~~~~~~~
