+++ app/new-bookshop/webapp/view/Object.view.xml
@@ -1,105 +1,144 @@
 <mvc:View controllerName="ns.newbookshop.controller.Object"
 	xmlns:mvc="sap.ui.core.mvc"
 	xmlns="sap.uxap"
-	xmlns:layout="sap.ui.layout"
-	xmlns:m="sap.m">
+	xmlns:m="sap.m"
+	xmlns:grid="sap.ui.table">
 
-	<ObjectPageLayout id="ObjectPageLayout" subSectionLayout="TitleOnLeft" showTitleInHeaderContent="true" upperCaseAnchorBar="false">
+	<ObjectPageLayout id="ObjectPageLayout" subSectionLayout="TitleOnLeft" upperCaseAnchorBar="false">
 		<headerTitle>
-			<ObjectPageDynamicHeaderTitle>
-				<expandedHeading>
-					<m:VBox>
-						<m:Title text="{title}" wrapping="true" class="sapUiSmallMarginBottom"/>
-						<m:Image src="{bookPictureURL}" width="80px"/>
-					</m:VBox>
-				</expandedHeading>
-				<snappedHeading>
-					<m:FlexBox fitContainer="true" alignItems="Center">
-						<m:Image src="{bookPictureURL}" class="sapUiSmallMarginEnd" width="100px"/>
-						<m:Title text="{title}" wrapping="true" />
-					</m:FlexBox>
-				</snappedHeading>
+			<ObjectPageHeader id="headerOP" objectTitle="{title}" objectSubtitle="{author/name}" objectImageURI="{bookPictureURL}" isObjectIconAlwaysVisible="true">
 				<actions>
-					<m:Button text="Edit" type="Emphasized"/>
-					<m:Button type="Transparent" text="Delete"/>
+					<m:Button type="Emphasized" text="{i18n>buttonEdit}" />
+					<m:Button type="Transparent" text="{i18n>buttonDelete}"/>
 				</actions>
-			</ObjectPageDynamicHeaderTitle>
+			</ObjectPageHeader>
 		</headerTitle>
 
 		<headerContent>
 			<m:FlexBox wrap="Wrap" fitContainer="true">
-				<layout:VerticalLayout class="sapUiSmallMarginBeginEnd">
-					<m:HBox>
-						<m:Label text="Price:" class="sapUiSmallMarginBeginEnd"/>
-						<m:ObjectNumber number="{
+				<m:HBox>
+					<m:Label text="{i18n>labelPrice}" class="sapUiSmallMarginBeginEnd" />
+					<m:ObjectNumber number="{
 												parts:[{path:'price'},{path:'currency_code'}],
 												type: 'sap.ui.model.type.Currency',
 												formatOptions: {showMeasure: false}
-											}" unit="{currency_code}" />
-					</m:HBox>
-				</layout:VerticalLayout>
+											}" unit="{currency_code}"/>
 
-				<layout:VerticalLayout class="sapUiSmallMarginBeginEnd">
-					<m:HBox>
-						<m:Label text="Stock:" class="sapUiSmallMarginBeginEnd"/>
-						<m:ObjectStatus class="sapUiSmallMarginBottom" text="{stock}" inverted="true" active="true" state="{
+					<m:Label text="{i18n>labelStock}" class="sapUiSmallMarginBeginEnd"/>
+					<m:ObjectStatus text="{stock}" inverted="true" active="false" state="{
                  								path: 'stock',
-                 								formatter: '.statusIndicatorColor'
+                 								formatter: '.formatter.statusIndicatorColor'
                 						}"/>
-					</m:HBox>
-				</layout:VerticalLayout>
-
+				</m:HBox>
 			</m:FlexBox>
 		</headerContent>
 
 		<sections>
-			<ObjectPageSection titleUppercase="false" title="Full description">
+			<ObjectPageSection titleUppercase="false" title="{i18n>titleFullDescription}">
 				<subSections>
-					<ObjectPageSubSection title="Description" titleUppercase="false">
-						<m:Text text="{descr}"/>
+					<ObjectPageSubSection title="{i18n>titleDescription}" titleUppercase="false">
+						<blocks>
+							<m:ExpandableText text="{descr}" overflowMode="InPlace" />
+						</blocks>
 					</ObjectPageSubSection>
 
-					<ObjectPageSubSection title="Genre" titleUppercase="false">
+					<ObjectPageSubSection title="{i18n>titleGenre}" titleUppercase="false">
 						<blocks>
 							<m:Text text="{genre}"/>
 						</blocks>
+
 					</ObjectPageSubSection>
 				</subSections>
 			</ObjectPageSection>
-			<ObjectPageSection titleUppercase="false" title="Rating">
+			<ObjectPageSection titleUppercase="false" title="{i18n>titleRating}">
 				<subSections>
 					<ObjectPageSubSection titleUppercase="false">
 						<blocks>
-							<m:RatingIndicator id="idRI" class="sapUiSmallMarginBottom" value="{rating}" />
+							<m:RatingIndicator id="idRI" class="sapUiSmallMarginBottom" value="{rating}" editable="false"/>
 						</blocks>
 					</ObjectPageSubSection>
 				</subSections>
 			</ObjectPageSection>
-			<ObjectPageSection titleUppercase="false" title="Orders">
+			<ObjectPageSection titleUppercase="false" title="{i18n>titleOrders}">
 				<subSections>
 					<ObjectPageSubSection titleUppercase="false">
 						<blocks>
-							<m:Table id="idOrdersTable" inset="false" items="{path:'orders',
-							parameters:{expand:'up_'}}">
+							<m:Table id="idOrdersTable" inset="false" items="{ 
+											 path:'orders',
+											  sorter : {
+													path : 'customerName',
+													group : true
+												}
+											  }">
 
 								<m:columns>
-									<m:Column width="12em">
-										<m:Text text="Customer name" />
+									<m:Column width="12em" >
+										<m:Text text="{i18n>tableNameColumnCustomerName}" />
 									</m:Column>
 									<m:Column>
-										<m:Text text="Delivery date" />
+										<m:Text text="{i18n>tableNameColumnDeliveryDate}" />
+									</m:Column>
+									<m:Column width="12em">
+										<m:Text text="{i18n>tableNameColumnPhoneNumber}" />
+									</m:Column>
+									<m:Column hAlign="End">
+										<m:Text text="{i18n>tableNameColumnCount}" />
 									</m:Column>
 
 								</m:columns>
 								<m:items>
 									<m:ColumnListItem vAlign="Middle">
 										<m:cells>
-											<m:Text text="{up_/customerName}" />
-											<m:Text text="{up_/deliveryDate}" />
+											<m:Text text="{customerName}" />
+											<m:Text text="{path: 'deliveryDate', type: 'sap.ui.model.type.Date' }" />
+											<m:Text text="{phoneNumber}" />
+											<m:Text text="{amount}" />
 										</m:cells>
 									</m:ColumnListItem>
 								</m:items>
 							</m:Table>
+
+							<!-- grid -->
+
+							<grid:Table alternateRowColors="true" 
+										rows="{orders}" 
+										id="idOrdersGridTable"
+										visibleRowCount="7" 
+										rowSelectionChange="enableDeleteButton" >
+								<grid:extension>
+									<m:OverflowToolbar style="Clear">
+										<m:Title id="title" text="{i18n>gridTableTitle}"/>
+										<m:ToolbarSpacer/>
+										<m:Button text="{i18n>buttonGridTableDelete}" enabled="{objectEnabled>/enabled}" press="onDeleteOrders" id="idDeleteButton"/>
+									</m:OverflowToolbar>
+								</grid:extension>
+								<grid:columns>
+									<grid:Column width="11rem" sortProperty="customerName" filterProperty="customerName">
+										<m:Label text="{{i18n>gridTableNameColumnCustomerName}" />
+										<grid:template>
+											<m:Text text="{customerName}" wrapping="false" />
+										</grid:template>
+									</grid:Column>
+									<grid:Column width="11rem" sortProperty="deliveryDate">
+										<m:Label text="{i18n>gridTableNameColumnDeliveryDate}" />
+										<grid:template>
+											<m:Text text="{path: 'deliveryDate', type: 'sap.ui.model.type.Date' }" wrapping="false" />
+										</grid:template>
+									</grid:Column>
+									<grid:Column width="11rem">
+										<m:Label text="{i18n>gridTableNameColumnPhoneNumber}" />
+										<grid:template>
+											<m:Text text="{phoneNumber}" wrapping="false" />
+										</grid:template>
+									</grid:Column>
+									<grid:Column width="11rem" sortProperty="amount" hAlign="End">
+										<m:Label text="{i18n>gridTableNameColumnCount}" />
+										<grid:template>
+											<m:Text text="{amount}" wrapping="false" />
+										</grid:template>
+									</grid:Column>
+								</grid:columns>
+							</grid:Table>
 						</blocks>
 					</ObjectPageSubSection>
 				</subSections>
