+++ db/data/sap.ui.bookshop-Orders-items.csv
@@ -1,20 +1,20 @@
 up__ID;book_ID;amount
 a5c19e98-7f1a-423b-a689-636f3adf6994;0c4510c4-5cee-4651-beb4-5f54b9c7d97c;1
-a5c19e98-7f1a-423b-a689-636f3adf6994;0c4510c4-5cee-4651-beb4-5f54b9c7d97c;1
-1fb7eee0-fbd7-4759-8609-e8e829f526b2;7aa8717f-07e7-484b-b987-8321faf38f2d;3
+a5c19e98-7f1a-423b-a689-636f3adf6994;7aa8717f-07e7-484b-b987-8321faf38f2d;3
+1fb7eee0-fbd7-4759-8609-e8e829f526b2;7aa8717f-07e7-484b-b987-8321faf38f2d;1
 41cd1b15-6d4e-4321-b3b4-2bae86acccd5;7aa8717f-07e7-484b-b987-8321faf38f2d;1
-5abaee8a-0b23-4515-9ecd-a780e6cd9fbb;0c4510c4-5cee-4651-beb4-5f54b9c7d97c;4
-a5c19e98-7f1a-423b-a689-636f3adf4444;7aa8717f-07e7-484b-b987-8321faf38f2d;1
-a5c19e98-7f1a-423b-a689-636f3adf4441;7aa8717f-07e7-484b-b987-8321faf38f2d;1
-a5c19e98-7f1a-423b-a689-636f3adf4442;7aa8717f-07e7-484b-b987-8321faf38f2d;1
-a5c19e98-7f1a-423b-a689-636f3adf4443;7aa8717f-07e7-484b-b987-8321faf38f2d;1
-a5c19e98-7f1a-423b-a689-636f3adf4445;7aa8717f-07e7-484b-b987-8321faf38f2d;1
-a5c19e98-7f1a-423b-a689-636f3adf4415;11f5072f-96b9-4d5e-9611-f04998e4c903;1
-a5c19e98-7f1a-423b-a689-636f3adf4425;11f5072f-96b9-4d5e-9611-f04998e4c903;1
+5abaee8a-0b23-4515-9ecd-a780e6cd9fbb;c11ae50c-bd4d-407b-b75d-3133c5c2ddd7;1
+a5c19e98-7f1a-423b-a689-636f3adf4444;0c4510c4-5cee-4651-beb4-5f54b9c7d97c;3
+a5c19e98-7f1a-423b-a689-636f3adf4441;11f5072f-96b9-4d5e-9611-f04998e4c903;1
+a5c19e98-7f1a-423b-a689-636f3adf4442;11f5072f-96b9-4d5e-9611-f04998e4c903;1
+a5c19e98-7f1a-423b-a689-636f3adf4443;11f5072f-96b9-4d5e-9611-f04998e4c903;1
+a5c19e98-7f1a-423b-a689-636f3adf4445;11f5072f-96b9-4d5e-9611-f04998e4c903;1
+a5c19e98-7f1a-423b-a689-636f3adf4415;11f5072f-96b9-4d5e-9611-f04998e4c903;4
+a5c19e98-7f1a-423b-a689-636f3adf4425;11f5072f-96b9-4d5e-9611-f04998e4c903;4
 a5c19e98-7f1a-423b-a689-636f3adf4435;11f5072f-96b9-4d5e-9611-f04998e4c903;1
-a5c19e98-7f1a-423b-a689-636f3adf4455;33671331-d6ec-4565-86f5-4b64ed8aa658;1
-a5c19e98-7f1a-423b-a689-636f3adf4465;33671331-d6ec-4565-86f5-4b64ed8aa658;1
-b5c19e98-7f1a-423b-a689-636f3adf4465;33671331-d6ec-4565-86f5-4b64ed8aa658;1
-c5c19e98-7f1a-423b-a689-636f3adf4465;33671331-d6ec-4565-86f5-4b64ed8aa658;1
-d5c19e98-7f1a-423b-a689-636f3adf4465;bbc37b13-d787-458a-8718-594779b19ae8;1
-7fc4794d-b5b8-48eb-bdec-243f0ea54484;bbc37b13-d787-458a-8718-594779b19ae8;1
+a5c19e98-7f1a-423b-a689-636f3adf4455;11f5072f-96b9-4d5e-9611-f04998e4c903;2
+a5c19e98-7f1a-423b-a689-636f3adf4465;11f5072f-96b9-4d5e-9611-f04998e4c903;1
+b5c19e98-7f1a-423b-a689-636f3adf4465;11f5072f-96b9-4d5e-9611-f04998e4c903;1
+c5c19e98-7f1a-423b-a689-636f3adf4465;11f5072f-96b9-4d5e-9611-f04998e4c903;2
+d5c19e98-7f1a-423b-a689-636f3adf4465;11f5072f-96b9-4d5e-9611-f04998e4c903;1
+7fc4794d-b5b8-48eb-bdec-243f0ea54484;11f5072f-96b9-4d5e-9611-f04998e4c903;1
